+++ CHANGES.md
@@ -7,6 +7,7 @@ Version 0.3.4
 -----------
 - added: `type: 'form'` config to `label`
 - added: `fill: 'screen'` config to `flexgrid`
+- added: `setValue()` to `label`
 - fixed: substr also matched by `showBatch`
 - fixed: `dropdown.master` refers to UI object instead of config
 - fixed: flexbox min-width bug with firefox